week15/pyproject.toml
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "black>=22.12.0",
]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Josh Hsieh", email = "jtr860830@gmail.com"},
]
dependencies = [
    "Pillow>=9.3.0",
]
requires-python = ">=3.8"
license = {text = "MIT"}
